refactor(actions): move body reading into base
also consolidating base Body interaction functions
