+++ base/body_test.go
@@ -1,0 +1,179 @@
+package base
+
+import (
+	"bytes"
+	"context"
+	"fmt"
+	"io/ioutil"
+	"net/http"
+	"net/http/httptest"
+	"testing"
+
+	"github.com/qri-io/dataset"
+	"github.com/qri-io/qfs"
+)
+
+func TestReadBody(t *testing.T) {
+	ctx := context.Background()
+	r := newTestRepo(t)
+	ref := addCitiesDataset(t, r)
+
+	ds, err := ReadDatasetPath(ctx, r, ref.String())
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	data, err := ReadBody(ds, dataset.JSONDataFormat, nil, 1, 1, false)
+	if err != nil {
+		t.Error(err.Error())
+	}
+	if !bytes.Equal(data, []byte(`[["new york",8500000,44.4,true]]`)) {
+		t.Errorf("byte response mismatch. got: %s", string(data))
+	}
+
+	if ds.BodyPath != "/map/QmcCcPTqmckdXLBwPQXxfyW2BbFcUT6gqv9oGeWDkrNTyD" {
+		t.Errorf("bodypath mismatch")
+	}
+}
+
+func TestDatasetPodBodyFile(t *testing.T) {
+	ctx := context.Background()
+	s := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		w.Write([]byte(`{"json":"data"}`))
+	}))
+	badS := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		w.WriteHeader(http.StatusInternalServerError)
+	}))
+
+	cases := []struct {
+		ds       *dataset.Dataset
+		filename string
+		fileLen  int
+		err      string
+	}{
+		// bad input produces no result
+		{&dataset.Dataset{}, "", 0, ""},
+
+		// inline data
+		{&dataset.Dataset{BodyBytes: []byte("a,b,c\n1,2,3")}, "", 0, "specifying bodyBytes requires format be specified in dataset.structure"},
+		{&dataset.Dataset{Structure: &dataset.Structure{Format: "csv"}, BodyBytes: []byte("a,b,c\n1,2,3")}, "body.csv", 11, ""},
+
+		// urlz
+		{&dataset.Dataset{BodyPath: "http://"}, "", 0, "fetching body url: Get http:: http: no Host in request URL"},
+		{&dataset.Dataset{BodyPath: fmt.Sprintf("%s/foobar.json", badS.URL)}, "", 0, "invalid status code fetching body url: 500"},
+		{&dataset.Dataset{BodyPath: fmt.Sprintf("%s/foobar.json", s.URL)}, "foobar.json", 15, ""},
+
+		// local filepaths
+		{&dataset.Dataset{BodyPath: "nope.cbor"}, "", 0, "body file: open nope.cbor: no such file or directory"},
+		{&dataset.Dataset{BodyPath: "nope.yaml"}, "", 0, "body file: open nope.yaml: no such file or directory"},
+		{&dataset.Dataset{BodyPath: "testdata/schools.cbor"}, "schools.cbor", 154, ""},
+		{&dataset.Dataset{BodyPath: "testdata/bad.yaml"}, "", 0, "converting yaml body to json: yaml: line 1: did not find expected '-' indicator"},
+		{&dataset.Dataset{BodyPath: "testdata/oh_hai.yaml"}, "oh_hai.json", 29, ""},
+	}
+
+	for i, c := range cases {
+		file, err := DatasetBodyFile(ctx, nil, c.ds)
+		if !(err == nil && c.err == "" || err != nil && err.Error() == c.err) {
+			t.Errorf("case %d error mismatch. expected: '%s', got: '%s'", i, c.err, err)
+			continue
+		}
+
+		if file == nil && c.filename != "" {
+			t.Errorf("case %d expected file", i)
+			continue
+		} else if c.filename == "" {
+			continue
+		}
+
+		if c.filename != file.FileName() {
+			t.Errorf("case %d filename mismatch. expected: '%s', got: '%s'", i, c.filename, file.FileName())
+		}
+		data, err := ioutil.ReadAll(file)
+		if err != nil {
+			t.Errorf("case %d error reading file: %s", i, err.Error())
+			continue
+		}
+		if c.fileLen != len(data) {
+			t.Errorf("case %d file length mismatch. expected: %d, got: %d", i, c.fileLen, len(data))
+		}
+
+		if err := file.Close(); err != nil {
+			t.Errorf("case %d error closing file: %s", i, err.Error())
+		}
+	}
+}
+
+func TestConvertBodyFormat(t *testing.T) {
+	jsonStructure := &dataset.Structure{Format: "json", Schema: dataset.BaseSchemaArray}
+	csvStructure := &dataset.Structure{Format: "csv", Schema: dataset.BaseSchemaArray}
+
+	// CSV -> JSON
+	body := qfs.NewMemfileBytes("", []byte("a,b,c"))
+	got, err := ConvertBodyFormat(body, csvStructure, jsonStructure)
+	if err != nil {
+		t.Error(err.Error())
+	}
+	data, err := ioutil.ReadAll(got)
+	if err != nil {
+		t.Fatal(err.Error())
+	}
+	if !bytes.Equal(data, []byte(`[["a","b","c"]]`)) {
+		t.Error(fmt.Errorf("converted body didn't match, got: %s", data))
+	}
+
+	// CSV -> JSON, multiple lines
+	body = qfs.NewMemfileBytes("", []byte("a,b,c\n\rd,e,f\n\rg,h,i"))
+	got, err = ConvertBodyFormat(body, csvStructure, jsonStructure)
+	if err != nil {
+		t.Fatal(err.Error())
+	}
+	data, err = ioutil.ReadAll(got)
+	if err != nil {
+		t.Fatal(err.Error())
+	}
+	if !bytes.Equal(data, []byte(`[["a","b","c"],["d","e","f"],["g","h","i"]]`)) {
+		t.Error(fmt.Errorf("converted body didn't match, got: %s", data))
+	}
+
+	// JSON -> CSV
+	body = qfs.NewMemfileBytes("", []byte(`[["a","b","c"]]`))
+	got, err = ConvertBodyFormat(body, jsonStructure, csvStructure)
+	if err != nil {
+		t.Fatal(err.Error())
+	}
+	data, err = ioutil.ReadAll(got)
+	if err != nil {
+		t.Fatal(err.Error())
+	}
+	if !bytes.Equal(data, []byte("a,b,c\n")) {
+		t.Error(fmt.Errorf("converted body didn't match, got: %s", data))
+	}
+
+	// CSV -> CSV
+	body = qfs.NewMemfileBytes("", []byte("a,b,c"))
+	got, err = ConvertBodyFormat(body, csvStructure, csvStructure)
+	if err != nil {
+		t.Fatal(err.Error())
+	}
+	data, err = ioutil.ReadAll(got)
+	if err != nil {
+		t.Fatal(err.Error())
+	}
+	if !bytes.Equal(data, []byte("a,b,c\n")) {
+		t.Error(fmt.Errorf("converted body didn't match, got: %s", data))
+	}
+
+	// JSON -> JSON
+	body = qfs.NewMemfileBytes("", []byte(`[["a","b","c"]]`))
+	got, err = ConvertBodyFormat(body, jsonStructure, jsonStructure)
+	if err != nil {
+		t.Fatal(err.Error())
+	}
+	data, err = ioutil.ReadAll(got)
+	if err != nil {
+		t.Fatal(err.Error())
+	}
+	if !bytes.Equal(data, []byte(`[["a","b","c"]]`)) {
+		t.Error(fmt.Errorf("converted body didn't match, got: %s", data))
+	}
+}
